Switched lat lng from decimal -> floats

+++ data/migrations/20210218201117_create_pins_table.js
@@ -6,8 +6,8 @@ exports.up = function(knex) {
         pins.string("name", 255)
         pins.string("title", 255).notNullable()
         pins.string("address", 255)
-        pins.decimal("lat").notNullable();
-        pins.decimal("lng").notNullable();
+        pins.float("lat").notNullable();
+        pins.float("lng").notNullable();
         pins.string("date").notNullable();
         pins.string("description")
         pins.string("image_url")
